requirements.txt
Django~=4.1.1
stripe~=4.2.0
PyJWT~=2.6.0
mintotp~=0.3.0
python-magic~=0.4.27